backend/requirements.txt
# Updated requirements.txt based on your working environment
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.11.2
pydantic-settings==2.10.1
pymongo==3.12.0
dnspython==2.4.2
python-jose==3.5.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.1.1
langchain-core==0.3.74
langchain==0.3.27
langchain-pinecone==0.2.11
langchain-google-genai==2.1.9
pinecone-client==6.0.0
langsmith==0.4.13
langchain-text-splitters==0.3.9

# Additional dependencies that might be needed
httpx==0.25.1
anyio==3.7.1
starlette==0.27.0